fix warning

+++ xlTetris/D2D10Renderer.cpp
@@ -153,10 +153,10 @@ void D2D10RenderContext::BeginDraw()
     float fDpiX = 0.0f, fDpiY = 0.0f;
     m_pRenderer->m_pD2DFactory->GetDesktopDpi(&fDpiX, &fDpiY);
 
-    rc.left = rc.left * fDpiX / 96.0f;
-    rc.top = rc.top * fDpiY / 96.0f;
-    rc.right = rc.right * fDpiX / 96.0f;
-    rc.bottom = rc.bottom * fDpiY / 96.0f;
+    rc.left = (int)(rc.left * fDpiX / 96.0f);
+    rc.top = (int)(rc.top * fDpiY / 96.0f);
+    rc.right = (int)(rc.right * fDpiX / 96.0f);
+    rc.bottom = (int)(rc.bottom * fDpiY / 96.0f);
 
     if (m_szLast.cx != rc.right - rc.left || m_szLast.cy != rc.bottom - rc.top)
     {
